Support for big intermediate xml files, Handle fbx with some no-ascii characters, changed fallback material to vivid red.
Former-commit-id: fc168abd9d7f756d4ab7f86f02146dba4f12baec

+++ converter.py
@@ -1,19 +1,23 @@
 import os
 import shutil
 import ase2xml
+import fbx2xml
 import fbx_info_extract
 import plybuilder
 import asexml2mitsubaxml
 
-filename = "scene"
+debug = False
+
+filename = "studio"
 ase2xml.ase2xmlconvert(filename)
+fbx2xml.transform(filename)
 fbx_info_extract.fbx_extract(filename)
 plybuilder.convert(filename)
 asexml2mitsubaxml.convert(filename)
 
-"""
-os.remove(filename+"_ase.xml")
-os.remove(filename+"_fbx.xml")
-shutil.rmtree("fbxinfos")
-shutil.rmtree("materials")
-"""
+if not debug :
+
+	os.remove(filename+"_ase.xml")
+	os.remove(filename+"_fbx.xml")
+	shutil.rmtree("fbxinfos")
+	shutil.rmtree("materials")
